backend/requirements.txt
# backend/requirements.txt

Django==4.2
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
djoser
python-dotenv
Pillow
dj-database-url # <--- Вот эта строка важна!
django-cors-headers
django-filter 
gunicorn